# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents:
    post:
      operationId: registerAgent
      tags: [agents]
      summary: Register a user-owned deterministic replay agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentRegistrationInput"
      responses:
        "201":
          description: Agent registered with a one-time credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRegistration"
        "400":
          description: Invalid agent registration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /agents/{agentId}:
    get:
      operationId: getAgent
      tags: [agents]
      parameters:
        - $ref: "#/components/parameters/AgentId"
        - $ref: "#/components/parameters/AgentCredential"
      responses:
        "200":
          description: Private agent details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "401":
          description: Missing or invalid agent credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /agents/{agentId}/wallet/challenge:
    post:
      operationId: createAgentWalletChallenge
      tags: [agents]
      parameters:
        - $ref: "#/components/parameters/AgentId"
        - $ref: "#/components/parameters/AgentCredential"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentWalletRequest"
      responses:
        "200":
          description: Short-lived wallet challenge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentWalletChallenge"
        "400":
          description: Invalid wallet address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid agent credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Agent credential changed while issuing challenge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /agents/{agentId}/wallet/bind:
    post:
      operationId: bindAgentWallet
      tags: [agents]
      parameters:
        - $ref: "#/components/parameters/AgentId"
        - $ref: "#/components/parameters/AgentCredential"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentWalletBinding"
      responses:
        "200":
          description: Wallet-bound agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "400":
          description: Invalid or expired wallet challenge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid agent credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Wallet challenge is no longer available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /agents/{agentId}/policy:
    patch:
      operationId: updateAgentPolicy
      tags: [agents]
      parameters:
        - $ref: "#/components/parameters/AgentId"
        - $ref: "#/components/parameters/AgentCredential"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentPolicyInput"
      responses:
        "200":
          description: Updated agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "400":
          description: Invalid or expired policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid agent credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Agent policy changed concurrently
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /agents/{agentId}/jobs:
    post:
      operationId: submitAgentJob
      tags: [agents]
      parameters:
        - $ref: "#/components/parameters/AgentId"
        - $ref: "#/components/parameters/AgentCredential"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentJobInput"
      responses:
        "201":
          description: Provider job queued for the agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderJob"
        "400":
          description: Invalid or disallowed job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid agent credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Agent or provider not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Agent or provider is not eligible
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /agents/{agentId}/events:
    get:
      operationId: listAgentEvents
      tags: [agents]
      parameters:
        - $ref: "#/components/parameters/AgentId"
        - $ref: "#/components/parameters/AgentCredential"
      responses:
        "200":
          description: Latest agent audit events
          content:
            application/json:
              schema:
                type: array
                maxItems: 100
                items:
                  $ref: "#/components/schemas/AgentEvent"
        "401":
          description: Missing or invalid agent credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /agents/{agentId}/credential/rotate:
    post:
      operationId: rotateAgentCredential
      tags: [agents]
      parameters:
        - $ref: "#/components/parameters/AgentId"
        - $ref: "#/components/parameters/AgentCredential"
      responses:
        "200":
          description: Replacement agent credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentCredentialRotation"
        "401":
          description: Missing or invalid agent credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Agent credential changed while rotating
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /agents/{agentId}/credential/revoke:
    post:
      operationId: revokeAgentCredential
      tags: [agents]
      parameters:
        - $ref: "#/components/parameters/AgentId"
        - $ref: "#/components/parameters/AgentCredential"
      responses:
        "204":
          description: Agent credential revoked
        "401":
          description: Missing or invalid agent credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Agent credential changed while revoking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
...
    AgentId:
      name: agentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AgentCredential:
      name: X-Isogate-Agent-Credential
      in: header
      required: true
      description: One-time credential returned by agent registration.
      schema:
        type: string
        minLength: 32
        maxLength: 128
...
  schemas:
...
    AgentRegistrationInput:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 64 }
        description: { type: string, maxLength: 240 }
...
    AgentPolicy:
      allOf:
        - $ref: "#/components/schemas/AgentPolicyInput"
    Agent:
      type: object
      additionalProperties: false
      required: [id, name, description, status, ownerWalletAddress, ownerWalletBoundAt, policy, policyVersion, policyHash, createdAt, updatedAt, lastSeenAt, revokedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 64 }
        description: { type: ["string", "null"], maxLength: 240 }
        status: { type: string, enum: [active, offline, revoked] }
        ownerWalletAddress: { type: ["string", "null"], pattern: "^0x[a-fA-F0-9]{40}$" }
        ownerWalletBoundAt: { type: ["string", "null"], format: date-time }
        policy: { $ref: "#/components/schemas/AgentPolicy" }
        policyVersion: { type: integer, minimum: 1 }
        policyHash: { type: string, pattern: "^[a-f0-9]{64}$" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        lastSeenAt: { type: string, format: date-time }
        revokedAt: { type: ["string", "null"], format: date-time }
    AgentRegistration:
      type: object
      additionalProperties: false
      required: [agent, credential]
      properties:
        agent: { $ref: "#/components/schemas/Agent" }
        credential: { type: string, minLength: 32, maxLength: 128 }
    AgentWalletRequest:
      type: object
      additionalProperties: false
      required: [walletAddress]
      properties:
        walletAddress: { type: string, pattern: "^0x[a-fA-F0-9]{40}$" }
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error